Move the auxiliary functions from the conftest.py of test_agentd to the agent.py module in wazuh_testing #1071
