+++ tests/integration/test_agentd/test_agent_auth_enrollment.py
@@ -2,6 +2,8 @@
 # Created by Wazuh, Inc. <info@wazuh.com>.
 # This program is free software; you can redistribute it and/or modify it under the terms of GPLv2
 
+import os
+import platform
 import pytest
 import subprocess
 import yaml
@@ -9,9 +11,9 @@ import yaml
 from wazuh_testing.tools.configuration import load_wazuh_configurations
 from wazuh_testing.tools.authd_sim import AuthdSimulator
 from wazuh_testing.tools.monitoring import QueueMonitor
-from conftest import AGENT_AUTH_BINARY_PATH, AUTHDPASS_PATH, DEFAULT_VALUES, \
+from wazuh_testing.agent import AGENT_AUTH_BINARY_PATH, AUTHDPASS_PATH, DEFAULT_VALUES, \
     SERVER_CERT_PATH, SERVER_KEY_PATH, AgentAuthParser, build_expected_request, \
-    check_client_keys_file, configure_enrollment, os, parse_configuration_string, platform
+    check_client_keys_file, configure_enrollment, parse_configuration_string
 
 # Marks
